requirements.txt
Flask==2.0.1
Werkzeug==2.0.3
flask-cors==3.0.10
pymysql==1.0.2
mysql-connector-python==8.0.28
cryptography
python-dotenv==0.19.0
requests==2.26.0
flask-swagger-ui==5.21.0
gunicorn==20.1.0